refactor(vendo): wire route table — apps (kill-list B4)
